refactor(StatUiModel): 국제화를 위해 name 프로퍼티를 @StringRes Int 로
