+++ android/app/src/main/java/poke/rogue/helper/presentation/dex/detail/PokemonDetailUiState.kt
@@ -8,14 +8,15 @@ import poke.rogue.helper.presentation.dex.model.EvolutionsUiModel
 import poke.rogue.helper.presentation.dex.model.PokemonBiomeUiModel
 import poke.rogue.helper.presentation.dex.model.PokemonDetailAbilityUiModel
 import poke.rogue.helper.presentation.dex.model.PokemonUiModel
-import poke.rogue.helper.presentation.dex.model.StatUiModel
+import poke.rogue.helper.presentation.dex.model.StatUiModel2
 import poke.rogue.helper.presentation.dex.model.toPokemonDetailUi
 import poke.rogue.helper.presentation.dex.model.toUi
+import poke.rogue.helper.presentation.dex.model.toUi2
 
 sealed interface PokemonDetailUiState {
     data class Success(
         val pokemon: PokemonUiModel,
-        val stats: List<StatUiModel>,
+        val stats: List<StatUiModel2>,
         val abilities: List<PokemonDetailAbilityUiModel>,
         val evolutions: EvolutionsUiModel,
         val skills: PokemonDetailSkills,
@@ -30,7 +31,7 @@ sealed interface PokemonDetailUiState {
 fun PokemonDetail.toUi(allBiomes: List<Biome>): PokemonDetailUiState.Success =
     PokemonDetailUiState.Success(
         pokemon = pokemon.toUi(),
-        stats = stats.map(Stat::toUi),
+        stats = stats.map(Stat::toUi2),
         abilities = abilities.toPokemonDetailUi(),
         evolutions = evolutions.toUi(),
         skills = skills,
@@ -42,7 +43,7 @@ fun PokemonDetail.toUi(allBiomes: List<Biome>): PokemonDetailUiState.Success =
 fun PokemonDetail.toUi(): PokemonDetailUiState.Success =
     PokemonDetailUiState.Success(
         pokemon = pokemon.toUi(),
-        stats = stats.map(Stat::toUi),
+        stats = stats.map(Stat::toUi2),
         abilities = abilities.toPokemonDetailUi(),
         evolutions = evolutions.toUi(),
         skills = skills,
