+++ android/app/src/main/java/poke/rogue/helper/presentation/dex/detail/stat/PokemonStatAdapter.kt
@@ -5,12 +5,12 @@ import android.view.ViewGroup
 import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView.ViewHolder
 import poke.rogue.helper.databinding.ItemStatBinding
-import poke.rogue.helper.presentation.dex.model.StatUiModel
+import poke.rogue.helper.presentation.dex.model.StatUiModel2
 import poke.rogue.helper.presentation.util.view.ItemDiffCallback
 import poke.rogue.helper.presentation.util.view.animation
 
 class PokemonStatAdapter :
-    ListAdapter<StatUiModel, PokemonStatAdapter.PokemonStatViewHolder>(statComparator) {
+    ListAdapter<StatUiModel2, PokemonStatAdapter.PokemonStatViewHolder>(statComparator) {
     override fun onCreateViewHolder(
         parent: ViewGroup,
         viewType: Int,
@@ -32,7 +32,7 @@ class PokemonStatAdapter :
 
     class PokemonStatViewHolder(private val binding: ItemStatBinding) :
         ViewHolder(binding.root) {
-        fun bind(stat: StatUiModel) {
+        fun bind(stat: StatUiModel2) {
             binding.stat = stat
             binding.progressBarStat.apply {
                 startAnimation(
@@ -50,7 +50,7 @@ class PokemonStatAdapter :
 
     companion object {
         private val statComparator =
-            ItemDiffCallback<StatUiModel>(
+            ItemDiffCallback<StatUiModel2>(
                 onItemsTheSame = { oldItem, newItem -> oldItem.name == newItem.name },
                 onContentsTheSame = { oldItem, newItem -> oldItem == newItem },
             )
